requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
boto3==1.9.126
botocore==1.12.126
cffi==1.12.2
colorama==0.4.1
cryptography==2.6.1
Django==2.1.7
django-cors-headers==2.5.2
django-filter==2.1.0
django-markdown-deux==1.0.5
django-model-utils==3.1.2
django-rest-knox==3.6.0
django-storages==1.7.1
djangorestframework==3.9.2
docutils==0.14
gunicorn==19.9.0
isort==4.3.16
jmespath==0.9.4
lazy-object-proxy==1.3.1
markdown2==2.3.7
mccabe==0.6.1
Pillow==5.4.1
psycopg2-binary==2.8.2
pycparser==2.19
pylint==2.3.1
pylint-django==2.0.6
pylint-plugin-utils==0.5
pyOpenSSL==19.0.0
python-dateutil==2.8.0
pytz==2018.9
s3transfer==0.2.0
six==1.12.0
typed-ast==1.3.1
urllib3==1.24.1
wrapt==1.11.1
